command: Use commands instead of functions
Replaces the various pip-app functions with a single function,
`pip-app`, that accepts commands just like `pip`.  One benefit of this
is that the user only has to remember a single command, `pip-app` and
pip-app itself has a help text describing each available command.

This change also introduces a command for listing installed applications
and their binaries.